+++ lib/spout/models/form.rb
@@ -1,0 +1,48 @@
+# {
+#   "id": "intake_questionnaire",
+#   "display_name": "Intake Questionnaire at Baseline Visit",
+#   "code_book": "Baseline-Visit-Intake-Questionnaire.pdf"
+# }
+
+require 'json'
+
+module Spout
+  module Models
+    class Form
+      attr_accessor :id, :display_name, :code_book
+      attr_accessor :errors
+
+      def initialize(file_name, dictionary_root)
+        @errors = []
+        @id     = file_name.to_s.gsub(/^(.*)\/|\.json$/, '').downcase
+        @folder = file_name.to_s.gsub(/^#{dictionary_root}\/forms\/|#{@id}\.json$/, '')
+
+        json = begin
+          JSON.parse(File.read(file_name))
+        rescue => e
+          form_name = file_name.to_s.gsub(/^(.*)\/|\.json$/, '').downcase
+          @errors << "Error Parsing #{form_name}.json: #{e.message}"
+          nil
+        end
+
+        if json and json.kind_of? Hash
+          %w( display_name code_book ).each do |method|
+            instance_variable_set("@#{method}", json[method])
+          end
+
+          @errors << "'id': #{json['id'].inspect} does not match filename #{@id.inspect}" if @id != json['id']
+        elsif json
+          @errors << "Form must be a valid hash in the following format: {\n\"id\": \"FORM_ID\",\n  \"display_name\": \"FORM DISPLAY NAME\",\n  \"code_book\": \"FORMPDF.pdf\"\n}"
+        end
+
+      end
+
+      def print
+        %w( id display_name code_book ).each do |method|
+          puts "#{"%13s" % method}: #{self.send(method).inspect}"
+        end
+      end
+
+    end
+  end
+end