+++ test/application/project_generator_test.rb
@@ -12,7 +12,7 @@ module ApplicationTests
     end
 
     def test_file_structure
-      assert_equal ['domains', 'test', 'variables', '.gitignore', '.ruby-version', '.spout.yml', '.travis.yml', 'Gemfile', 'Rakefile', '.', '..'].sort, Dir.entries(app_path).sort
+      assert_equal ['domains', 'forms', 'test', 'variables', '.gitignore', '.ruby-version', '.spout.yml', '.travis.yml', 'Gemfile', 'Rakefile', '.', '..'].sort, Dir.entries(app_path).sort
     end
   end
 end